Exposing kernel gram APIs
